Refactor posflag's default flag value behaviour.
- posflag.Provider() now takes an optional Koanf instance to see if the
the flags defined have been set from other providers, for instance,
a config file. If they are not, then the default values of the flags
are merged. If they do exist, the flag values are not merged but only
the values that have been explicity set in the command line are merged.
- Change the test package to 'koanf_test' to avoid import cycles.