feat: add optional delim to structs provider
This commit adds a new feature to the structs provider which takes an
optional delimiter. In case a delimiter is provided, the structs
provider will apply koanf/maps.Unflatten which will unflatten the map
returned by the fatih/structs package.

To maintain backward compatibility this is added as an optional
behavior instead.

Closes: https://github.com/knadh/koanf/issues/122