Fix StringMap and StringsMap
When calling `StringMap`, or `StringsMap`, and the `ko.Get()` returns
the value as expected, e.g. `map[string]string` for `StringMap`, the
type assertion for `map[string]interface{}` does not match. This
causes it to return an empty map even when there are actual values.

This commit addresses this by simply returning a copy if the type
assertion matches what we're supposed to return.