Remove buggy `mapstructure.StringToSliceHookFunc(",")`.
The `StringToSliceHookFunc(",")` hook is supposed to decode a delim
separated string to a slice of strings. However, the hook only receives
`reflect.Kind` for the target field to determine its type. While it
should be a `[]string`, it can only tell if it's a slice,
thus triggering it for any slice type and failing. This breaks
the unmarshalling of strings to `[]byte` fields.